Delete redundant PhoneBook

+++ src/main/java/com/dins/demo/data/ContactRepository.java
@@ -4,5 +4,5 @@ import com.dins.demo.domain.Contact;
 import org.springframework.data.repository.CrudRepository;
 
 public interface ContactRepository
-        extends CrudRepository<Contact, Long> {
+        extends CrudRepository<Contact, Integer> {
 }
