+++ src/main/java/com/dins/demo/data/UserRepository.java
@@ -4,5 +4,5 @@ import com.dins.demo.domain.User;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository
-        extends CrudRepository<User, Long> {
+        extends CrudRepository<User, Integer> {
 }
